+++ apps/ced-portal-be/CHANGELOG.md
@@ -1,5 +1,21 @@
 # ced-portal-be
 
+## 0.0.15
+
+### Patch Changes
+
+- c92a9f0: Abstracted service name into infra
+- Updated dependencies [c92a9f0]
+  - @pagopa/io-core-adapter-tracing@0.0.2
+
+## 0.0.14
+
+### Patch Changes
+
+- 15c464d: Added azure tracing adapter with hooks for fastify.
+- Updated dependencies [15c464d]
+  - @pagopa/io-core-adapter-tracing@0.0.1
+
 ## 0.0.13
 
 ### Patch Changes